fix tests add stub for gitcharm
