+++ tests/test_deployer.py
@@ -497,55 +497,7 @@ class DeployerTests(unittest.TestCase):
         pass
 
 
-class CharmCacheTest(unittest.TestCase):
-    def test_init(self):
-        c = CharmCache('mytestcharm')
-        self.assertEqual(c.test_charm, 'mytestcharm')
 
-    @patch('amulet.deployer.get_charm')
-    def test_getitem_service(self, get_charm):
-        c = CharmCache('mytestcharm')
-        charm = c['myservice']
-        self.assertEqual(charm, get_charm.return_value)
-        get_charm.assert_called_once_with('myservice', series='precise')
-
-        get_charm.reset_mock()
-        charm2 = c['myservice']
-        self.assertEqual(charm, charm2)
-        self.assertFalse(get_charm.called)
-
-    @patch('amulet.deployer.get_charm')
-    def test_getitem_testcharm(self, get_charm):
-        c = CharmCache('mytestcharm')
-        charm = c['mytestcharm']
-        self.assertEqual(charm, get_charm.return_value)
-        get_charm.assert_called_once_with(os.getcwd(), series='precise')
-
-    @patch('amulet.deployer.get_charm')
-    def test_fetch_service(self, get_charm):
-        c = CharmCache('mytestcharm')
-        charm = c.fetch('myservice')
-        self.assertEqual(charm, get_charm.return_value)
-        get_charm.assert_called_once_with('myservice', series='precise')
-
-        get_charm.reset_mock()
-        charm2 = c['myservice']
-        self.assertEqual(charm, charm2)
-        self.assertFalse(get_charm.called)
-
-    @patch('amulet.deployer.get_charm')
-    def test_fetch_charm(self, get_charm):
-        c = CharmCache('mytestcharm')
-        charm = c.fetch('myservice', 'anothercharm')
-        self.assertEqual(charm, get_charm.return_value)
-        get_charm.assert_called_once_with('anothercharm', series='precise')
-
-    @patch('amulet.deployer.get_charm')
-    def test_fetch_testcharm(self, get_charm):
-        c = CharmCache('mytestcharm')
-        charm = c.fetch('myservice', 'mytestcharm')
-        self.assertEqual(charm, get_charm.return_value)
-        get_charm.assert_called_once_with(os.getcwd(), series='precise')
 
 
 class GetCharmNameTest(unittest.TestCase):
